feat!: switch RolePolicy resource to ccapi provider
This switches the `RolePolicy` resource from the `aws` provider to the
`aws-native` provider. The `aws` provider has some eventual consistency
issues that the `ccapi` provider does not. With the `ccapi` provider you
can `dependOn` the policy resource and it actually waits for it to have
been deployed everywhere.

BREAKING CHANGE: `RolePolicy` resource not uses the `ccapi` provider